+++ modules/processing/static.py
@@ -115,6 +115,13 @@ try:
 except ImportError as e:
     HAVE_PEEPDF = False
 
+try:
+    HAVE_XLM_DEOBF = True
+    from XLMMacroDeobfuscator.deobfuscator import process_file as XLMMacroDeobf
+except ImportError:
+    print("Missed dependey XLMMacroDeobfuscator: pip3 install git+https://github.com/DissectMalware/XLMMacroDeobfuscator.git")
+    HAVE_XLM_DEOBF = False
+
 log = logging.getLogger(__name__)
 processing_conf = Config("processing")
 
@@ -1657,6 +1664,20 @@ class Office(object):
                 metares["DocumentType"] = indicator.name
             if indicator.name == "PowerPoint Presentation" and indicator.value == True:
                 metares["DocumentType"] = indicator.name
+
+        if HAVE_XLM_DEOBF and processing_conf.xlsdeobf.enabled:
+            xlm_kwargs = {
+                'file': filepath,
+                'noninteractive': True,
+                'extract_only': False,
+                'no_ms_excel': True,
+                'start_with_shell': False,
+                'return_deobfuscated': True,
+                'day': 0,
+            }
+            deofuscated_xlm = XLMMacroDeobf(**xlm_kwargs)
+            if deofuscated_xlm:
+                results["office"]["XLMMacroDeobfuscator"] = deofuscated_xlm
         return results
 
     def run(self):